Add Helper module to CreateOne and One classes
