+++ lib/storage_tables/attached.rb
@@ -10,6 +10,6 @@ module StorageTables
   end
 end
 
+require "storage_tables/attachable/changes"
 require "storage_tables/attachable/model"
 require "storage_tables/attachable/one"
-require "storage_tables/attachable/changes"
